refactor : calorie 타입을 double로 변경, FoodResponse에서 calorie를 반올림하여 int로 반환

+++ backend/src/main/java/com/fit/nufit/food/domain/Food.java
@@ -30,9 +30,9 @@ public class Food extends BaseEntity {
     private FoodType type;
 
     @Column(name = "food_calorie")
-    private int calorie;
+    private double calorie;
 
-    public Food(String name, int amount, String brand, FoodType foodType, int calorie) {
+    public Food(String name, int amount, String brand, FoodType foodType, double calorie) {
         this.name = name;
         this.brand = brand;
         this.amount = amount;
